Implement property initialization framework with data source handling
- Added DataSourceInitializer to manage data source initialization and ensure proper lifecycle handling.
- Introduced ObjectInitializationService for executing IAsyncInitializer during test execution.
- Created ObjectRegistrationService for handling object registration and property injection during the registration phase.
- Developed PropertyDataResolver to resolve data from property data sources and manage initialization.
- Established property initialization strategies (IPropertyInitializationStrategy) for nested and reflection-based properties.
- Implemented PropertyInitializationOrchestrator to coordinate property initialization processes.
- Defined PropertyInitializationPipeline to manage the execution flow of property initialization.
- Enhanced PropertyInjectionService to utilize the new initialization framework for injecting properties into objects.
