+++ TUnit.Engine/Framework/TUnitServiceProvider.cs
@@ -7,8 +7,6 @@ using Microsoft.Testing.Platform.Messages;
 using Microsoft.Testing.Platform.Requests;
 using Microsoft.Testing.Platform.Services;
 using TUnit.Core;
-using TUnit.Core.DataSources;
-using TUnit.Core.Initialization;
 using TUnit.Core.Interfaces;
 using TUnit.Core.Tracking;
 using TUnit.Engine.Building;
