+++ TUnit.Engine/Services/PropertyInitialization/Strategies/ReflectionPropertyStrategy.cs
@@ -1,39 +1,49 @@
-using TUnit.Core.DataSources;
-using TUnit.Core.Initialization;
+using System.Diagnostics.CodeAnalysis;
+using System.Runtime.CompilerServices;
+using TUnit.Core;
+using TUnit.Core.PropertyInjection.Initialization;
 
-namespace TUnit.Core.PropertyInjection.Initialization.Strategies;
+namespace TUnit.Engine.Services.PropertyInitialization;
 
 /// <summary>
-/// Strategy for initializing properties using source-generated metadata.
-/// Optimized for AOT and performance.
+/// Strategy for initializing properties using reflection.
+/// Used when source generation is not available.
 /// </summary>
-internal sealed class SourceGeneratedPropertyStrategy : IPropertyInitializationStrategy
+internal sealed class ReflectionPropertyStrategy : IPropertyInitializationStrategy
 {
     private readonly DataSourceInitializer _dataSourceInitializer;
     private readonly ObjectRegistrationService _objectRegistrationService;
 
-    public SourceGeneratedPropertyStrategy(DataSourceInitializer dataSourceInitializer, ObjectRegistrationService objectRegistrationService)
+    public ReflectionPropertyStrategy(DataSourceInitializer dataSourceInitializer, ObjectRegistrationService objectRegistrationService)
     {
         _dataSourceInitializer = dataSourceInitializer ?? throw new System.ArgumentNullException(nameof(dataSourceInitializer));
         _objectRegistrationService = objectRegistrationService ?? throw new System.ArgumentNullException(nameof(objectRegistrationService));
     }
     /// <summary>
-    /// Determines if this strategy can handle source-generated properties.
+    /// Determines if this strategy can handle reflection-based properties.
     /// </summary>
     public bool CanHandle(PropertyInitializationContext context)
     {
-        return context.SourceGeneratedMetadata != null && SourceRegistrar.IsEnabled;
+        return context.PropertyInfo != null && context.DataSource != null && !SourceRegistrar.IsEnabled;
     }
 
     /// <summary>
-    /// Initializes a property using source-generated metadata.
+    /// Initializes a property using reflection.
     /// </summary>
+    [UnconditionalSuppressMessage("Trimming", "IL2072", Justification = "Reflection mode support")]
     public async Task InitializePropertyAsync(PropertyInitializationContext context)
     {
-        if (context.SourceGeneratedMetadata == null)
+#if NET
+        if (!RuntimeFeature.IsDynamicCodeSupported)
+        {
+            throw new Exception("Using TUnit Reflection mechanisms isn't supported in AOT mode");
+        }
+#endif
+
+        if (context.PropertyInfo == null || context.DataSource == null)
         {
             return;
-        }
+        };
 
         object? resolvedValue = null;
 
@@ -45,6 +55,7 @@ internal sealed class SourceGeneratedPropertyStrategy : IPropertyInitializationS
         }
         else
         {
+            // Step 1: Resolve data from the data source (execution-time resolution)
             resolvedValue = await PropertyDataResolver.ResolvePropertyDataAsync(context, _dataSourceInitializer, _objectRegistrationService);
             if (resolvedValue == null)
             {
@@ -54,8 +65,11 @@ internal sealed class SourceGeneratedPropertyStrategy : IPropertyInitializationS
             context.ResolvedValue = resolvedValue;
         }
 
-        context.SourceGeneratedMetadata.SetProperty(context.Instance, resolvedValue);
+        // Step 3: Set the property value
+        // The value has already been initialized by PropertyDataResolver if needed
+        context.PropertySetter(context.Instance, resolvedValue);
 
+        // Step 4: Add to test context tracking (if not already there)
         if (context.TestContext != null && !context.TestContext.TestDetails.TestClassInjectedPropertyArguments.ContainsKey(context.PropertyName))
         {
             context.TestContext.TestDetails.TestClassInjectedPropertyArguments[context.PropertyName] = resolvedValue;
